requirements.txt
backports-abc==0.5
bkcharts==0.2
bokeh==0.12.6
certifi==2017.4.17
chardet==3.0.4
cycler==0.10.0
functools32==3.2.3.post2
futures==3.1.1
idna==2.5
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.0
pandas==0.20.2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.18.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
tornado==4.5.1
urllib3==1.21.1
virtualenv==15.1.0
